↪️ move tracer func to logger package
