+++ explorer/scan/discovery.go
@@ -12,12 +12,12 @@ import (
 	"go.mondoo.com/cnquery/v11/cli/config"
 	"go.mondoo.com/cnquery/v11/cli/execruntime"
 	"go.mondoo.com/cnquery/v11/llx"
+	"go.mondoo.com/cnquery/v11/logger"
 	"go.mondoo.com/cnquery/v11/providers"
 	inventory "go.mondoo.com/cnquery/v11/providers-sdk/v1/inventory"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/inventory/manager"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/plugin"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/upstream"
-	"go.mondoo.com/cnquery/v11/providers-sdk/v1/util/tracer"
 )
 
 type AssetWithRuntime struct {
@@ -153,7 +153,7 @@ func DiscoverAssets(ctx context.Context, inv *inventory.Inventory, upstream *ups
 }
 
 func discoverAssets(rootAssetWithRuntime *AssetWithRuntime, resolvedRootAsset *inventory.Asset, discoveredAssets *DiscoveredAssets, runtimeLabels map[string]string, upstream *upstream.UpstreamConfig, recording llx.Recording) {
-	defer tracer.FuncDur(time.Now(), "explorer.discoverAssets")
+	defer logger.FuncDur(time.Now(), "explorer.discoverAssets")
 
 	// It is possible that we did not discover any assets under the root asset. In that case the inventory
 	// would be nil and we can return
