+++ providers/github/provider/provider.go
@@ -11,10 +11,10 @@ import (
 	"time"
 
 	"go.mondoo.com/cnquery/v11/llx"
+	"go.mondoo.com/cnquery/v11/logger"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/inventory"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/plugin"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/upstream"
-	"go.mondoo.com/cnquery/v11/providers-sdk/v1/util/tracer"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/vault"
 	"go.mondoo.com/cnquery/v11/providers/github/connection"
 	"go.mondoo.com/cnquery/v11/providers/github/resources"
@@ -189,7 +189,7 @@ func (s *Service) connect(req *plugin.ConnectReq, callback plugin.ProviderCallba
 }
 
 func (s *Service) detect(asset *inventory.Asset, conn *connection.GithubConnection) error {
-	defer tracer.FuncDur(time.Now(), "provider.github.service.detect")
+	defer logger.FuncDur(time.Now(), "provider.github.service.detect")
 
 	conf := asset.Connections[0]
 	asset.Name = conf.Host
@@ -215,7 +215,7 @@ func (s *Service) detect(asset *inventory.Asset, conn *connection.GithubConnecti
 }
 
 func (s *Service) discover(conn *connection.GithubConnection) (*inventory.Inventory, error) {
-	defer tracer.FuncDur(time.Now(), "provider.github.service.discover")
+	defer logger.FuncDur(time.Now(), "provider.github.service.discover")
 
 	conf := conn.Asset().Connections[0]
 	if conf.Discover == nil {
