+++ providers/github/resources/discovery.go
@@ -12,9 +12,9 @@ import (
 	"github.com/google/go-github/v62/github"
 	"github.com/rs/zerolog/log"
 	"go.mondoo.com/cnquery/v11/llx"
+	"go.mondoo.com/cnquery/v11/logger"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/inventory"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/plugin"
-	"go.mondoo.com/cnquery/v11/providers-sdk/v1/util/tracer"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/vault"
 	"go.mondoo.com/cnquery/v11/providers/github/connection"
 	"go.mondoo.com/cnquery/v11/utils/stringx"
@@ -51,7 +51,7 @@ func handleTargets(targets []string) []string {
 }
 
 func discover(runtime *plugin.Runtime, targets []string) ([]*inventory.Asset, error) {
-	defer tracer.FuncDur(time.Now(), "provider.github.discover")
+	defer logger.FuncDur(time.Now(), "provider.github.discover")
 
 	conn := runtime.Connection.(*connection.GithubConnection)
 	conf := conn.Asset().Connections[0]
