+++ providers/github/resources/github_repo.go
@@ -15,9 +15,9 @@ import (
 	"github.com/google/go-github/v62/github"
 	"github.com/rs/zerolog/log"
 	"go.mondoo.com/cnquery/v11/llx"
+	"go.mondoo.com/cnquery/v11/logger"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/plugin"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/util/convert"
-	"go.mondoo.com/cnquery/v11/providers-sdk/v1/util/tracer"
 	"go.mondoo.com/cnquery/v11/providers/github/connection"
 	"go.mondoo.com/cnquery/v11/types"
 )
@@ -134,7 +134,7 @@ func initGithubRepository(runtime *plugin.Runtime, args map[string]*llx.RawData)
 	if len(args) > 2 {
 		return args, nil, nil
 	}
-	defer tracer.FuncDur(time.Now(), "provider.github.initGithubRepository")
+	defer logger.FuncDur(time.Now(), "provider.github.initGithubRepository")
 
 	conn := runtime.Connection.(*connection.GithubConnection)
