+++ providers/github/resources/github_user.go
@@ -12,8 +12,8 @@ import (
 	"github.com/cockroachdb/errors"
 	"github.com/google/go-github/v62/github"
 	"go.mondoo.com/cnquery/v11/llx"
+	"go.mondoo.com/cnquery/v11/logger"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/plugin"
-	"go.mondoo.com/cnquery/v11/providers-sdk/v1/util/tracer"
 	"go.mondoo.com/cnquery/v11/providers/github/connection"
 	"go.mondoo.com/cnquery/v11/types"
 	"go.mondoo.com/cnquery/v11/utils/stringx"
@@ -37,7 +37,7 @@ func initGithubUser(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[
 		return args, nil, nil
 	}
 
-	defer tracer.FuncDur(time.Now(), "provider.github.initGithubUser")
+	defer logger.FuncDur(time.Now(), "provider.github.initGithubUser")
 	conn := runtime.Connection.(*connection.GithubConnection)
 
 	userLogin := ""
